pyproject.toml
[tool.poetry]
name = "webcam-int8"
version = "0.0.1"
description = "Human-segmentation-int8-model"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyvirtualcam = "^0.11.1"
opencv-python = "^4.9.0.80"
onnxruntime-gpu = "^1.17.1"
numba = "^0.59.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
